+++ flow/stub/jsonModules.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+type JsonModuleType = { [key: string]: any }
+const emptyJSONModule: JsonModuleType = {}
+export default emptyJSONModule