+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog #
 
+## v1.0.1 ##
+
+* React-hot-reload on client side and live-reload on server side in combination with Browsersync for testing on external devices.
+* Webpack bundle for client and server side.
+* Webpack-dev-server instead of dev middleware.
+* Isomorphic server-side rendering.
+* Code splitting and long-term caching of assets.
+* Easy import of Bootstrap variables and mixins
+
 ## v1.0.0 ##
 
 First release